reorganize algos
