+++ src/faebryk/core/solver/defaultsolver.py
@@ -17,9 +17,15 @@ from faebryk.core.parameter import (
     ParameterOperatable,
     Predicate,
 )
-from faebryk.core.solver import analytical, canonical, literal_folding
 from faebryk.core.solver.mutator import REPR_MAP, Mutator
 from faebryk.core.solver.solver import LOG_PICK_SOLVE, Solver
+from faebryk.core.solver.symbolic import (
+    canonical,
+    expression_groups,
+    expression_wise,
+    pure_literal,
+    structural,
+)
 from faebryk.core.solver.utils import (
     ALLOW_PARTIAL_STATE,
     MAX_ITERATIONS_HEURISTIC,
@@ -68,31 +74,31 @@ class DefaultSolver(Solver):
             canonical.alias_predicates_to_true,
         ],
         iterative=[
-            analytical.check_literal_contradiction,
-            analytical.remove_unconstrained,
-            analytical.convert_operable_aliased_to_single_into_literal,
-            analytical.resolve_alias_classes,
-            analytical.distribute_literals_across_alias_classes,
-            analytical.remove_congruent_expressions,
-            analytical.convert_inequality_with_literal_to_subset,
-            analytical.compress_associative,
-            analytical.reflexive_predicates,
-            analytical.idempotent_deduplicate,
-            analytical.idempotent_unpack,
-            analytical.involutory_fold,
-            analytical.unary_identity_unpack,
-            literal_folding.fold_pure_literal_expressions,
-            *literal_folding.fold_algorithms,
-            analytical.merge_intersect_subsets,
-            analytical.predicate_flat_terminate,
-            analytical.predicate_unconstrained_operands_deduce,
-            analytical.predicate_terminated_is_true,
-            analytical.empty_set,
-            analytical.transitive_subset,
-            analytical.isolate_lone_params,
-            analytical.remove_empty_graphs,
-            analytical.uncorrelated_alias_fold,
-            analytical.upper_estimation_of_expressions_with_subsets,
+            structural.check_literal_contradiction,
+            structural.remove_unconstrained,
+            structural.convert_operable_aliased_to_single_into_literal,
+            structural.resolve_alias_classes,
+            structural.distribute_literals_across_alias_classes,
+            structural.remove_congruent_expressions,
+            structural.convert_inequality_with_literal_to_subset,
+            expression_groups.associative_flatten,
+            expression_groups.reflexive_predicates,
+            expression_groups.idempotent_deduplicate,
+            expression_groups.idempotent_unpack,
+            expression_groups.involutory_fold,
+            expression_groups.unary_identity_unpack,
+            pure_literal.fold_pure_literal_expressions,
+            *expression_wise.fold_algorithms,
+            structural.merge_intersect_subsets,
+            structural.predicate_flat_terminate,
+            structural.predicate_unconstrained_operands_deduce,
+            structural.predicate_terminated_is_true,
+            structural.empty_set,
+            structural.transitive_subset,
+            structural.isolate_lone_params,
+            structural.remove_empty_graphs,
+            structural.uncorrelated_alias_fold,
+            structural.upper_estimation_of_expressions_with_subsets,
         ],
     )
