+++ src/faebryk/core/solver/symbolic/canonical.py
