+++ test/core/solver/test_solver.py
@@ -1506,7 +1506,9 @@ def test_fold_correlated():
     ],
 )
 def test_exec_pure_literal_expressions(op: type[CanonicalExpression], lits, expected):
-    from faebryk.core.solver.literal_folding import _exec_pure_literal_expressions
+    from faebryk.core.solver.symbolic.expression_wise import (
+        _exec_pure_literal_expressions,
+    )
     from faebryk.core.solver.utils import make_lit
 
     lits_converted = list(map(make_lit, lits))
